gitconfig追加

+++ install.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
  
+if [ ! -e _gitconfig ]; then
+    echo "You must modify _gitconfig.example as _gitconfig before execute this script"
+    exit 1
+fi
+
 ln -sf ~/dotfiles/.vimrc ~/.vimrc
 ln -sf ~/dotfiles/.vim ~/.vim
 ln -sf ~/dotfiles/.zshrc ~/.zshrc
